indexwriter is part of core datastore API, should not have been refactored
