+++ core/ingest/src/main/java/mil/nga/giat/geowave/core/ingest/local/IngestTask.java
@@ -11,9 +11,9 @@ import java.util.concurrent.TimeUnit;
 import mil.nga.giat.geowave.core.index.ByteArrayId;
 import mil.nga.giat.geowave.core.ingest.GeoWaveData;
 import mil.nga.giat.geowave.core.store.AdapterToIndexMapping;
+import mil.nga.giat.geowave.core.store.IndexWriter;
 import mil.nga.giat.geowave.core.store.adapter.WritableDataAdapter;
 import mil.nga.giat.geowave.core.store.index.PrimaryIndex;
-import mil.nga.giat.geowave.core.store.index.writer.IndexWriter;
 import mil.nga.giat.geowave.core.store.util.DataStoreUtils;
 
 import org.slf4j.Logger;
