+++ core/store/src/main/java/mil/nga/giat/geowave/core/store/index/writer/IndexCompositeWriter.java
@@ -6,6 +6,7 @@ import java.util.Arrays;
 import java.util.List;
 
 import mil.nga.giat.geowave.core.index.ByteArrayId;
+import mil.nga.giat.geowave.core.store.IndexWriter;
 import mil.nga.giat.geowave.core.store.data.VisibilityWriter;
 import mil.nga.giat.geowave.core.store.index.PrimaryIndex;
