+++ examples/src/main/java/mil/nga/giat/geowave/examples/ingest/SimpleIngestProducerConsumer.java
@@ -12,8 +12,8 @@ import org.opengis.feature.simple.SimpleFeatureType;
 
 import mil.nga.giat.geowave.adapter.vector.GeotoolsFeatureDataAdapter;
 import mil.nga.giat.geowave.core.store.DataStore;
+import mil.nga.giat.geowave.core.store.IndexWriter;
 import mil.nga.giat.geowave.core.store.index.PrimaryIndex;
-import mil.nga.giat.geowave.core.store.index.writer.IndexWriter;
 import mil.nga.giat.geowave.datastore.accumulo.BasicAccumuloOperations;
 
 public class SimpleIngestProducerConsumer extends
